# openapi.json
{"openapi":"3.1.0","info":{"title":"AI Slides API","version":"1.0.0"},"paths":{"/api/v1/presentations/":{"post":{"summary":"Create Presentation","description":"Create a new presentation:\n1. Generate presentation JSON from topic\n2. Store in database\n3. Generate requested file type (pptx/pdf) with UUID as filename\n4. Return UUID","operationId":"create_presentation_api_v1_presentations__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresentationCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresentationCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Presentations","description":"List all presentations with pagination","operationId":"list_presentations_api_v1_presentations__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/presentations/{presentation_id}":{"get":{"summary":"Get Presentation","description":"Get presentation by UUID\nReturns: UUID and main_topic from database","operationId":"get_presentation_api_v1_presentations__presentation_id__get","parameters":[{"name":"presentation_id","in":"path","required":true,"schema":{"type":"string","title":"Presentation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresentationGetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Presentation","description":"Delete presentation by UUID\nRemoves entry from database and deletes associated files","operationId":"delete_presentation_api_v1_presentations__presentation_id__delete","parameters":[{"name":"presentation_id","in":"path","required":true,"schema":{"type":"string","title":"Presentation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Slide","description":"Update a specific slide in the presentation\n- Takes slide_number and slide_content\n- Updates the presentation JSON in database\n- Regenerates files with same UUID\n- Returns UUID","operationId":"update_slide_api_v1_presentations__presentation_id__put","parameters":[{"name":"presentation_id","in":"path","required":true,"schema":{"type":"string","title":"Presentation Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SlideUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SlideUpdateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PresentationCreate":{"properties":{"main_topic":{"type":"string","minLength":1,"title":"Main Topic","description":"Main topic/prompt for presentation"},"file_type":{"type":"string","enum":["pptx","pdf"],"title":"File Type","description":"Type of file to generate"}},"type":"object","required":["main_topic","file_type"],"title":"PresentationCreate","description":"POST request - create presentation and generate file"},"PresentationCreateResponse":{"properties":{"id":{"type":"string","title":"Id","description":"UUID of the presentation"}},"type":"object","required":["id"],"title":"PresentationCreateResponse","description":"Response with UUID after creation"},"PresentationGetResponse":{"properties":{"id":{"type":"string","title":"Id","description":"UUID of the presentation"},"main_topic":{"type":"string","title":"Main Topic","description":"Main topic of the presentation"}},"type":"object","required":["id","main_topic"],"title":"PresentationGetResponse","description":"GET response - presentation data from database"},"SlideUpdate":{"properties":{"slide_number":{"type":"integer","minimum":1.0,"title":"Slide Number","description":"Slide number (1-based)"},"slide_content":{"type":"string","minLength":1,"title":"Slide Content","description":"New content for the slide"}},"type":"object","required":["slide_number","slide_content"],"title":"SlideUpdate","description":"PUT request - update specific slide"},"SlideUpdateResponse":{"properties":{"id":{"type":"string","title":"Id","description":"UUID of the presentation"}},"type":"object","required":["id"],"title":"SlideUpdateResponse","description":"Response after slide update"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}